Move check_region_obligations_and_report_errors to TypeErrCtxt
